app/requirements.txt
fastapi
uvicorn
langchain
openai
faiss-cpu
pandas
python-dotenv
